feat: create add todo
all calls for /user checks if the user is logged in and checks if they are a user using middleware. Checks task validation. Adds todo to Todo model, and pushes the task._id to todo attribute for userMetadata model. Renamed the Task and completedTask model to Todo and completedTodo.

fixes #14