refactor(community): change field communityId to communityTag
Changed the communityId field in Community model to communityTag to avoid confusion with _id while coding the functionalities. Changes every code which refers to the prior communityId to communityTag, and not changed the communityId which refered to the _id. Removed checking for existing communityTag while creating a new community from communityServices as it is already checked by validator.

Fixes #45