feat: create delete community by community admin
Checks if community exists, and checks if user is the community admin. Throws error if any of these doesn't match. The community will be deleted only if the user is the community admin. The params is communityId, and userId is retrieved from req.auth. Proper error handling is implemented.

Closed #44