feat: add moderator functionality
Added the functionality for a community admin to add a new moderator. Only the community admin can perform this function, checks if the moderatorId is a valid userId. Throws error if the user that is to be added as moderator is not a member in community. Adds to moderator array only if the user is not a moderator, else it throws an error.

Resolve #65