feat: create leave community functionality
created a funcitonality for members to leave the community. Members and moderators can leave, but throws an error when community admin tries to leave. Community admin can only leave after assigning the community admin role to another member. Checks if the user is a member or not before removing them from the community. Removes the userId from moderator list if included. Changed previousUser field to previousMembers, the userId is pushed to this list while leaving if not included already.

resolves #69